+++ apps/storefront/src/shared/service/b2b/graphql/login.ts
@@ -25,3 +25,42 @@ export const getBCGraphqlToken = (data: Partial<ApiTokenConfig>): Promise<string
     variables: { storeFrontTokenData: convertObjectOrArrayKeysToCamel(data) },
   }).then((res) => res.storeFrontToken.token);
 };
+
+interface BCPermission {
+  code: string;
+  permissionLevel: number;
+}
+
+interface BCAuthorizationInput {
+  bcToken: string;
+  channelId: number;
+}
+
+interface BCAuthorizationResponse {
+  authorization?: {
+    result?: {
+      permissions: BCPermission[];
+    };
+  };
+  errors?: Array<{ message: string }>;
+}
+
+const GET_BC_AUTHORIZATION = `mutation BCAuthorization($authData: UserAuthType!) {
+  authorization(authData: $authData) {
+    result {
+      permissions {
+        code
+        permissionLevel
+      }
+    }
+  }
+}`;
+
+export async function bcAuthorization(
+  authData: BCAuthorizationInput,
+): Promise<BCAuthorizationResponse> {
+  return B3Request.graphqlB2B<BCAuthorizationResponse>({
+    query: GET_BC_AUTHORIZATION,
+    variables: { authData },
+  });
+}
