+++ apps/storefront/src/shared/service/b2b/index.ts
@@ -34,7 +34,7 @@ import {
   superAdminCompanies,
   superAdminEndMasquerade,
 } from './graphql/global';
-import { getBCGraphqlToken } from './graphql/login';
+import { bcAuthorization, getBCGraphqlToken } from './graphql/login';
 import {
   getB2BAllOrders,
   getB2BOrderDetails,
@@ -118,6 +118,7 @@ export { getOrderedProducts } from './graphql/quickOrder';
 export {
   addProductToBcShoppingList,
   addProductToShoppingList,
+  bcAuthorization,
   B2BProductsBulkUploadCSV,
   quoteCheckout,
   BcProductsBulkUploadCSV,
