+++ apps/storefront/src/shared/service/bc/graphql/login.ts
@@ -86,42 +86,3 @@ export const bcLogoutLogin = () =>
   storefrontGQLRequest({
     query: logoutLogin(),
   });
-
-interface BCPermission {
-  code: string;
-  permissionLevel: number;
-}
-
-interface BCAuthorizationInput {
-  bcToken: string;
-  channelId: number;
-}
-
-interface BCAuthorizationResponse {
-  authorization?: {
-    result?: {
-      permissions: BCPermission[];
-    };
-  };
-  errors?: Array<{ message: string }>;
-}
-
-const GET_BC_AUTHORIZATION = `mutation BCAuthorization($authData: UserAuthType!) {
-  authorization(authData: $authData) {
-    result {
-      permissions {
-        code
-        permissionLevel
-      }
-    }
-  }
-}`;
-
-export async function bcAuthorization(
-  authData: BCAuthorizationInput,
-): Promise<BCAuthorizationResponse> {
-  return B3Request.graphqlB2B<BCAuthorizationResponse>({
-    query: GET_BC_AUTHORIZATION,
-    variables: { authData },
-  });
-}
